Updated to native Julia libraries
I'll keep a legacy record on FrFK-Pycall.jl of the notebook for future reference where I was using the definitions of Hyperfunctions in terms of PyCall wrapper and mpmath libray.

The updated notebook FrFK.jl had the new implementations in terms of SpecialFunctions.jl and HypergeometricFunctions.jl. A substancial speedup is observed.